pyproject.toml
[build-system]
requires = ["setuptools>=61", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "data_copilot"
description = "The data_copilot web application"
readme = "README.rst"
requires-python = ">=3.10"
keywords = ["data_copilot"]
authors = [
  {name = "Modulos AG", email = "contact@modulos.ai" }
]
classifiers = [
  "Development Status :: 2 - Pre-Alpha",
  "Intended Audience :: Developers",
  "Natural Language :: English",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
]
dynamic = ["version"]

dependencies = [
  "aiohttp>=3.8.3",
  "alembic>=1.10.2",
  "bcrypt>=4.0.1",
  "build>=0.10.0",
  "celery>=5.2.7",
  "click>=8.1.3",
  "colorama>=0.4.6",
  "email-validator>=1.3.0",
  "fastapi-sso>=0.6.4",
  "fastapi>=0.86.0",
  "Flask>=2.3.2",
  "fsspec>=2023.1.0",
  "langchain>=0.0.179",
  "numpy>=1.23.2",
  "openai>=0.27",
  "openpyxl>=3.1.1",
  "pandas>=1.5.3",
  "passlib>=1.7.4",
  "psycopg2-binary>=2.9.5",
  "pydantic>=1.10.4",
  "python-dotenv>=0.21.1",
  "python-jose>=3.3.0",
  "python-multipart>=0.0.5",
  "redis>=4.4.2",
  "simplejson>=3.18.3",
  "sqlalchemy>=2.0.4",
  "tabulate==0.9.0",
  "uvicorn>=0.20.0",
  "watchdog>=2.2.1",
  "xlrd>=2.0.1"
]

[project.scripts]
  data-copilot = "data_copilot.main:main"

[project.urls]
"Homepage" = "https://github.com/Modulos/data_copilot/"
"Bug Reports" = "https://github.com/Modulos/data_copilot/issues"
"Source" = "https://github.com/Modulos/data_copilot/"

[tool.setuptools.packages.find]
include = ["data_copilot"]

[tool.setuptools_scm]
write_to = "data_copilot/_version.py"

[tool.pytest.ini_options]
minversion = "3.0"
testpaths = ["tests"]
addopts = [
    "--import-mode=importlib",
]
